Initial ssh subcommand
The ssh subcommand is meant to be used from an ssh authorized_keys
forced command.  It parses SSH_ORIGINAL_COMMAND, sanitizes the
arguments, and runs recognized commands.

The initial recognized command is rsync, which is only allowed in sender
mode, i.e. copying from the backup source and not writing to it.  A few
dangerous arguments are filtered, but this could definitely stand to be
locked down more.  Currently usable to back up files readable by the
backup account.  Needs to implement sudo support to back up other files.