Add basic sudo subcommand
The sudo subcommand can run the rsync server with the same validation as
the ssh command.  It doesn't do PATH lookups because it assumes that the
incoming ssh subcommand did that before invoking sudo.  Making ssh
decide to call sudo will follow after adding a way to look up the
current host in the config.