Refactor CliArgs
We want to pass the global args to ssh so that it can regenerate correct
command-line args for calling sudo.  The current CliArgs is sitting in
main.rs and also includes other things that ssh shouldn't need to know
about.  Refactor the global args into a separate struct that gets
flattened with structopt and move the whole setup into a separate file.

This also requires making most of the commands take themselves as a
reference instead of consuming their object, since otherwise we end up
with a partially moved-from CliArgs object.