Open log file more carefully
When running with sudo, the log file will be opened as root.  As long as
the initial forced ssh command is passed an absolute path that doesn't
have any components writable by an untrusted user (potentially including
the backup user itself), this should be mostly safe.  doppelback
initially opens the log file when running as the normal user, so
pointing the arg at a system file will fail before sudo is invoked.

By blocking symlinks when opening the file, we can prevent the backup
user from racing against itself to get sudo to overwrite some other
file.  This still relies on being unable to manipulate parents in the
path, but that's hopefully an easier bar to clear.  This can eventually
be fixed further with openat2 once that's available everywhere.

Eventually it would be even better to pass the log as an open fd through
the sudo invocation and not open the log as root at all.  However, this
requires some extra sudo setup or a parent-child relationship instead of
exec.  In the meantime, we can make this a little safer by
canonicalizing the path, passing it as an absolute path, and ignoring
symlinks.