Add sudo support to ssh command
ssh now accepts the original global command-line arguments, the path to
the running binary, and a backup host config.  For rsync, it finds the
path being read in SSH_ORIGINAL_COMMAND, uses that to look up the backup
source in the host config, and then builds a sudo command if needed.