Make config required
All of the existing commands except `sudo` require a config.  When
`sudo` is called, it will receive the config that was passed in for the
generating command, so it effectively always gets it as well.  We can
print a nicer error message by just making `--config` required.  If a
future command is added that doesn't require a config after all, it can
be switched back and handled similarly to `--host`.