Add remote config-test command
In addition to all the local tests, make `config-test` ssh into each
backup host and remotely run `doppelback config-test` for each source.
When run remotely, `config-test` will check that the specified source
directory is found in the config and corresponds to a real directory.
This also double-checks the sudo setup if any sources request root.

Bump version to 0.3.0 since the new remote command will not work
correctly if the remote host is running an older version.