Rework path checks for unreadable source dirs
If the backup source path is unreadable by a non-root user, doppelback
cannot canonicalize it before elevating with sudo.  Instead of failing
for this case, change the initial check to only require absolute paths.

Move the full canonicalization to the step after sudo has been run,
right before running the rsync server.  At this point, if doppelback
can't access the source dir then rsync is going to fail anyway.

For the non-root case, this leaves open the possibility of running the
rsync server from a path containing symlinks.  However, the source path
has to match an entry from the config file and doppelback doesn't have
any extra privileges in this case, so it seems reasonable to rely on the
user to enter the path they intend.